FlutterWordGame
Both players are asked to choose how many letter words they want to enter and enter the word. While entering the word, there is a 1-minute period. After entering the word and confirming it, the remaining time is added to their last score.
They have the right to guess as many as the number of letters selected.
After each guess, the letters in the correct position are painted green, while the letter in the wrong position in the word is painted yellow.
Scoring is green square*10 + yellow square*5.